spec/requirements.txt
asgiref==3.6.0
concurrent-log-handler==0.9.22
coverage==7.2.3
deepmerge==1.1.0
Django==4.1.8
django-auth-ldap==4.3.0
django-cleanup==6.0.0
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
gunicorn
jira==3.4.1
ldap3==2.9.1
mock==5.0.2
mssql-django==1.3
pandas==2.0.1
portalocker==2.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyodbc==4.0.39
PyPDF2==2.10.5
python-ldap
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
